+++ .devcontainer/test.sh
@@ -1,0 +1,6 @@
+#!/bin/sh -e
+
+. ~/.bashrc
+
+cd /tmp/checkout
+make checks